Added functionality.
Triangle, generates a triangle based on the cursor and the drawPanel. Needs to get the center of the triangle.

-Circle

+++ src/figures/GeometricFigure.java
@@ -6,8 +6,8 @@ abstract public class GeometricFigure{
 
     private int x, y;
 
-    public abstract float getPerimeter();
-    public abstract float getArea();
+    public abstract double getPerimeter();
+    public abstract double getArea();
     public abstract void draw(Graphics g);
 
     GeometricFigure(int x, int y) {
